+++ dms-20250414/alibabacloud_dms20250414/models/__init__.py
@@ -220,6 +220,7 @@ from ._delete_airflow_response_body import DeleteAirflowResponseBodyRootResponse
 from ._delete_airflow_response_body import DeleteAirflowResponseBodyRoot
 from ._delete_file_upload_response_body import DeleteFileUploadResponseBodyData
 from ._describe_custom_agent_response_body import DescribeCustomAgentResponseBodyDataExecutionConfig
+from ._describe_custom_agent_response_body import DescribeCustomAgentResponseBodyDataKnowledgeConfigList
 from ._describe_custom_agent_response_body import DescribeCustomAgentResponseBodyData
 from ._describe_data_agent_session_response_body import DescribeDataAgentSessionResponseBodyDataChatHistoryLocations
 from ._describe_data_agent_session_response_body import DescribeDataAgentSessionResponseBodyDataSessionConfig
@@ -233,6 +234,7 @@ from ._get_notebook_task_status_response_body import GetNotebookTaskStatusRespon
 from ._list_airflows_response_body import ListAirflowsResponseBodyRootList
 from ._list_airflows_response_body import ListAirflowsResponseBodyRoot
 from ._list_custom_agent_response_body import ListCustomAgentResponseBodyDataContentExecutionConfig
+from ._list_custom_agent_response_body import ListCustomAgentResponseBodyDataContentKnowledgeConfigList
 from ._list_custom_agent_response_body import ListCustomAgentResponseBodyDataContent
 from ._list_custom_agent_response_body import ListCustomAgentResponseBodyData
 from ._list_data_agent_workspace_response_body import ListDataAgentWorkspaceResponseBodyDataContent
@@ -466,6 +468,7 @@ __all__ = [
     DeleteAirflowResponseBodyRoot,
     DeleteFileUploadResponseBodyData,
     DescribeCustomAgentResponseBodyDataExecutionConfig,
+    DescribeCustomAgentResponseBodyDataKnowledgeConfigList,
     DescribeCustomAgentResponseBodyData,
     DescribeDataAgentSessionResponseBodyDataChatHistoryLocations,
     DescribeDataAgentSessionResponseBodyDataSessionConfig,
@@ -479,6 +482,7 @@ __all__ = [
     ListAirflowsResponseBodyRootList,
     ListAirflowsResponseBodyRoot,
     ListCustomAgentResponseBodyDataContentExecutionConfig,
+    ListCustomAgentResponseBodyDataContentKnowledgeConfigList,
     ListCustomAgentResponseBodyDataContent,
     ListCustomAgentResponseBodyData,
     ListDataAgentWorkspaceResponseBodyDataContent,